# api/openapi/management-v1.yaml
openapi: 3.1.0
info:
  title: Laneway Management API
  version: 1.0.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  summary: Versioned administrator management and browser-session contract.
  description: >-
    The current v1 administrator surface. It intentionally documents bounded
    list snapshots and cursor-paginated audit streams alongside current
    mutation semantics without claiming totals, entity tags, preconditions, or
    idempotency keys that are not yet implemented. JSON request objects are
    strict and reject unknown fields.
...
paths:
...
  /v1/admin/service-principals:
    post:
      operationId: createServicePrincipal
      tags: [Automation]
      summary: Create a scoped automation principal.
      description: >-
        Owner session or root automation. Permissions and network visibility are
        explicit; no human owner role is inherited.
      security:
        - administratorSession: []
          administratorCsrfCookie: []
          administratorCsrfHeader: []
        - rootBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml#/components/schemas/CreateServicePrincipalRequest'
      responses:
        '201':
          $ref: './components.yaml#/components/responses/ServicePrincipalCreated'
        default:
          $ref: './components.yaml#/components/responses/Error'
    get:
      operationId: listServicePrincipals
      tags: [Automation]
      summary: Read a bounded service-principal snapshot.
      description: >-
        Enabled principals sort first and are capped at 100, so the default
        100-record inventory always includes every enabled automation identity.
        A larger limit may include disabled history.
      security:
        - administratorSession: []
        - rootBearer: []
      parameters:
        - $ref: './components.yaml#/components/parameters/Limit'
      responses:
        '200':
          $ref: './components.yaml#/components/responses/ServicePrincipals'
        default:
          $ref: './components.yaml#/components/responses/Error'
  /v1/admin/service-principals/{principal_id}/disable:
    post:
      operationId: disableServicePrincipal
      tags: [Automation]
      summary: Disable an automation principal and revoke its tokens.
      description: >-
        Owner session or root automation. Disabling is irreversible in v1 and
        atomically revokes every unrevoked token belonging to the principal.
      security:
        - administratorSession: []
          administratorCsrfCookie: []
          administratorCsrfHeader: []
        - rootBearer: []
      parameters:
        - $ref: './components.yaml#/components/parameters/PrincipalID'
      responses:
        '204':
          $ref: './components.yaml#/components/responses/SessionNoContent'
        default:
          $ref: './components.yaml#/components/responses/Error'
  /v1/admin/service-principals/{principal_id}/tokens:
    post:
      operationId: issueServiceAccessToken
      tags: [Automation]
      summary: Issue a service access token.
      description: The bearer is returned once and only its digest is persisted.
      security:
        - administratorSession: []
          administratorCsrfCookie: []
          administratorCsrfHeader: []
        - rootBearer: []
      parameters:
        - $ref: './components.yaml#/components/parameters/PrincipalID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml#/components/schemas/IssueServiceAccessTokenRequest'
      responses:
        '201':
          $ref: './components.yaml#/components/responses/ServiceAccessTokenIssued'
        default:
          $ref: './components.yaml#/components/responses/Error'
    get:
      operationId: listServiceAccessTokens
      tags: [Automation]
      summary: Read safe token metadata for one service principal.
      description: >-
        Live tokens sort first. At most 100 unrevoked tokens may belong to one
        principal, so the default 100-record inventory always includes every
        bearer that can still authorize a request. A larger limit may include
        revoked history.
      security:
        - administratorSession: []
        - rootBearer: []
      parameters:
        - $ref: './components.yaml#/components/parameters/PrincipalID'
        - $ref: './components.yaml#/components/parameters/Limit'
      responses:
        '200':
          $ref: './components.yaml#/components/responses/ServiceAccessTokens'
        default:
          $ref: './components.yaml#/components/responses/Error'
  /v1/admin/service-access-tokens/{token_id}/revoke:
    post:
      operationId: revokeServiceAccessToken
      tags: [Automation]
      summary: Revoke a service access token immediately.
      security:
        - administratorSession: []
          administratorCsrfCookie: []
          administratorCsrfHeader: []
        - rootBearer: []
      parameters:
        - $ref: './components.yaml#/components/parameters/TokenID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml#/components/schemas/ServiceAccessTokenRevocationRequest'
      responses:
        '204':
          $ref: './components.yaml#/components/responses/SessionNoContent'
        default:
          $ref: './components.yaml#/components/responses/Error'
...
components:
  securitySchemes:
    administratorSession:
      $ref: './components.yaml#/components/securitySchemes/administratorSession'
    administratorCsrfCookie:
      $ref: './components.yaml#/components/securitySchemes/administratorCsrfCookie'
    administratorCsrfHeader:
      $ref: './components.yaml#/components/securitySchemes/administratorCsrfHeader'
    rootBearer:
      $ref: './components.yaml#/components/securitySchemes/rootBearer'
    serviceAccessToken:
      $ref: './components.yaml#/components/securitySchemes/serviceAccessToken'